Error and Usability improvements
Improved the Error-Handling by better storing and propagating the core Error that
caused a chain of Errors to occur.
Then I also changed the way Channel-Permissions are managed. Now the active channels
are not accessible by the @everyone-Role, to not annoy everyone else that is not playing
in the current Round, and only the W-Dead-Role, the Mods and the Users
with the corresponding Werewolf-Role are given Permission to access the Channel. At the end
of the Game, all this is reset again to the Default-Permissions, allowing everyone to read
the Channel again.