Made a lot more adjustments
Introduced a simple Trait to abstract over the concrete BotContext, which for now simply exposes some internal
stuff again but now helps to only pass around a dyn BotContext instead of the concrete Context. For now this does
not in turn provide any meaningful improvements or fixes anything, but instead will help to slowly migrate features
over to that trait to allow for mocking these parts of the system for testing purposes.