First Changes
Made some more code improvements for better readability and reusability.
Then also made the first switch by not allowing users to enter as Moderators themselves but rather only adding
users as mods when they have the correct Role on game start. However there is still more work to do in this
aspect as well because the code is currently just panicing if it cant find a Role with the correct Name, which
is obviously not a good solution