'Use' StorageBackend
Added the StorageBackend to the Bot as well as a provisional `list-roles` command to test the current State of
handling Roles.
Right now it always returns an empty List of Roles and can't actually save any Roles.

In the Future we need to define a generic Role, that contains all the Data needed for a Role, and then also add
the ability to serialize and deserialize a Role to be able to store it. This may also force us to redesign or at least
adapt other parts of the Codebase that interact with the 'Role System'.
We will also need a way to "cache" Roles as it would be very inefficient and kind of dumb to always load all the Roles
and configuration from Scratch before doing anything that requires them, in the Real World we only need to load them
on startup as well as keep it up to date whenever we update/add/remove a Role from the Guild